Refactor out oauth code
